Make TicketCard issue and PR chips clickable
Extract LinkChip into its own file so the ticket board rows can reuse
the same clickable capsule the session header already uses. Replace
the static purple PR badge and the plain issue-number text in
TicketCard with gold LinkChips that open the issue and PR URLs in the
default browser.
